renaming useGameData file
